pyproject.toml
[tool.poetry]
name = "blogwriter"
version = "0.1.0"
description = "Framework modulare per la generazione automatizzata di articoli scientifici tramite CrewAI."
authors = ["acimarelli <alessandro.cimarelli@icloud.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.2.3"
numpy = "^2.2.6"
matplotlib = "^3.10.3"
scikit-learn = "^1.6.1"
jupyterlab = "^4.4.2"
ipykernel = "^6.29.5"
crewai = "^0.120.1"
crewai-tools = "^0.45.0"
pyyaml = "^6.0.2"
python-dotenv = "^1.1.0"
langchain-ollama = "^0.3.3"
pydantic = "^2.7"
streamlit = "^1.35.0"
graphviz = "^0.20.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
black = "^24.4.2"
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
